[refactor] updated create_tx example to create_simple_tx example, added readme, and updated the code with comments
