+++ docs/examples/broadcaster/broadcaster.md
@@ -1,0 +1,8 @@
+# Example: Building a Custom Transaction Broadcast Client
+This guide walks through the necessary steps for building a custom transaction broadcast client.
+
+## Overview
+A transaction broadcast client is a crucial component in any Bitcoin SV application, allowing it to communicate with the Bitcoin SV network. Implementing a transaction broadcaster can be accomplished using the clearly defined Broadcast interface.
+
+Our task will be to create a broadcaster that connects with the What's on Chain service. This broadcaster is particularly designed for browser applications and utilizes the standard Fetch API for HTTP communications with the relevant API endpoints.
+
